fix: renames access control list component with imports to match changes made in DMSS
